Fewer panics and more exceptions on more access violations
Ideally the simulator should never panic, but there were no
propagating of the errors from the lowest level of memory access.
This improves the situation a great deal, but unsurprisingly it had a
large ripple effect and you can't just give store_phys_u8() the
ability to return an error without changing *all* the memory access
functions.

This is definitely not the last memory change.