Uop cache
This is a change I have building up to for a long time.  We refactor
the instruction execution into fetch, decode, register fetch, and
execute.  This enables us to cache the uop and amortize the fetch and
decode.  The register fetch is generic and what remains is invoking
the execution for the uop.

This already leads to a measurably perf improvement, but no tuning has
been performed yet and we currently only cache individual uops.